Bot_Practica/requirements.txt
MySQL-python==1.2.5
web.py==0.38
python-telegram-bot==5.3.0
